Move DevUp.Persistence project to DevUp.Infrastructure
