+++ backend/DevUp/DevUp.Infrastructure/DevUp.Infrastructure.csproj
@@ -2,8 +2,12 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+    <ImplicitUsings>disable</ImplicitUsings>
+    <Nullable>disable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DevUp.Domain\DevUp.Domain.csproj" />
+  </ItemGroup>
 
 </Project>
