--- backend/DevUp/DevUp.Persistence/DevUp.Persistence.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>disable</ImplicitUsings>
-    <Nullable>disable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\DevUp.Domain\DevUp.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
