+++ backend/DevUp/DevUp.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevUp.Api", "DevUp.Api\DevU
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevUp.Domain", "DevUp.Domain\DevUp.Domain.csproj", "{2488F971-7F5A-47A1-91BF-FAE754B8B5DF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevUp.Persistence", "DevUp.Persistence\DevUp.Persistence.csproj", "{B199B8EE-E9DE-4F47-95C8-FBEC40372FF4}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevUp.Infrastructure", "DevUp.Infrastructure\DevUp.Infrastructure.csproj", "{5A5FC462-08B5-4E34-A27B-C9AEF9A4814A}"
 EndProject
 Global
@@ -25,10 +23,6 @@ Global
 		{2488F971-7F5A-47A1-91BF-FAE754B8B5DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2488F971-7F5A-47A1-91BF-FAE754B8B5DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2488F971-7F5A-47A1-91BF-FAE754B8B5DF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B199B8EE-E9DE-4F47-95C8-FBEC40372FF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B199B8EE-E9DE-4F47-95C8-FBEC40372FF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B199B8EE-E9DE-4F47-95C8-FBEC40372FF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B199B8EE-E9DE-4F47-95C8-FBEC40372FF4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5A5FC462-08B5-4E34-A27B-C9AEF9A4814A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5A5FC462-08B5-4E34-A27B-C9AEF9A4814A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5A5FC462-08B5-4E34-A27B-C9AEF9A4814A}.Release|Any CPU.ActiveCfg = Release|Any CPU
